app/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage==2.7.0
torch==1.13.0
numpy==1.23.4
opencv-contrib-python-headless==4.7.0.68